requirements.txt
aiohttp>=3.9
websockets>=12.0
python-dotenv>=1.0
pytest>=8.0
